fix: Remove absolute paths in source checksum logic
When we see if a dataset is out of date, we identify its ingredients and
checksum them all. Then we combine the checksums, sorting by filename.
However the filenames were absolute paths, which vary by machine.

Whilst this shouldn't change the sort order, still we make these paths
relative in order to remove one source of difference between different
machines this code is run on.